$ cargo new hello_cargo --bin
"We passed the --bin argument to cargo new because our goal is to make an executable application, as opposed to a library. Executables are binary executable files often called just binaries. We’ve given hello_cargo as the name for our project, and Cargo creates its files in a directory of the same name that we can then go into."