Generate a random number with Rand crate
"Knowing which traits to use and which functions and methods to call from a crate isn’t something that you’ll just know. Instructions for using a crate are in each crate’s documentation. Another neat feature of Cargo is that you can run the cargo doc --open command that will build documentation provided by all of your dependencies locally and open it in your browser. If you’re interested in other functionality in the rand crate, for example, run cargo doc --open and click rand in the sidebar on the left."